Add `-version` option to get just the version
Adds a `-version` or `--version` option to print just the version
of the program for bitcoind, bitcoin-cli and bitcoin-qt.

Also make it that `-help` can be used to display the help (as well as
existing `--help`). Up to now, `-help` was the only option that didn't
work with either one or two dashes.

Conflicts:
	src/bitcoin-cli.cpp
	src/bitcoind.cpp
	src/qt/utilitydialog.cpp